Substitute @PACKAGE_NAME@ with the package's name.
This is done in source files and in the phar stub in the same way as for
@PACKAGE_VERSION@.

This makes it possible to reuse the same set of files across several projects
(eg. one stub for many different projects or modules of the same project)
while still having the correct (specific) output file for each project,
because the substitution occurs at package creation time (during the make
or package command).